add python pkgs install division

+++ postgresql/balance/provision/env_provision_lvirt.sh
@@ -42,10 +42,17 @@ sudo apt-get install htop -y
 sudo apt-get install -y gcc make perl
 sudo apt-get install linux-[5-6].*-generic -y
 sudo apt-get install linux-[5-6].*-lowlatency -y
-sudo apt-get install -y python3-pip
-python3 -m pip install --upgrade pip
-python3 -m pip install psycopg2-binary
 dpkg -s ntpsec &> /dev/null || sudo apt-get install ntpsec -y
+sudo apt-get install -y python3-pip
+
+if test "$(grep -E '1.8.*' /etc/astra_version)"; then
+    python3 -m pip install --upgrade pip --break-system-packages
+    python3 -m pip install psycopg2-binary --break-system-packages
+elif test "$(grep -E '1.7.*' /etc/astra_version)"; then
+    python3 -m pip install --upgrade pip
+    python3 -m pip install psycopg2-binary
+fi
+
 
 
 main_user=vagrant
